Moved the contract state event reader as a method to the TransactionsReader. (required to ease parallel stories development)
