+++ ledger/participant-integration-api/src/main/scala/platform/store/dao/MeteredLedgerDao.scala
@@ -74,9 +74,6 @@ private[platform] class MeteredLedgerReadDao(ledgerDao: LedgerReadDao, metrics: 
 
   override def transactionsReader: LedgerDaoTransactionsReader = ledgerDao.transactionsReader
 
-  override def contractStateEventsReader: LedgerDaoContractStateEventsReader =
-    ledgerDao.contractStateEventsReader
-
   override def lookupKey(key: GlobalKey, forParties: Set[Party])(implicit
       loggingContext: LoggingContext
   ): Future[Option[Value.ContractId]] =
